applesingle: Added custom "streaming iterator."
Still working on an easy-to-use interface with the current language
capabilities, but it is definitely capable of iterating over segments of the
input stream which are either:
 - Raw data for a resource/data fork as std::io::Read
 - Raw data for an unsupported passthrough data format as std::io::Read
 - Interpreted small chunks of data such as finder info, dates, filename,
   comment, etc.

This is all done without ever going backwards in the stream (or without a native
forward-seek) since it may be from a pipe or the network.

There may be a more advanced alternative of this in the future for archives
small enough to be held in memory or seekable/mmap()-ed files.